% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{status_db}
\alias{status_db}
\title{persistent connection}
\usage{
status_db()
}
\description{
persistent connection
}
\examples{
\dontrun{
# should not do this, it would raise a warning
status_db() |> tbl("basic") |> head() # leaves connection open
status_db() |> tbl("BcChkERR") |> head()
}
con = status_db()
con |> tbl("basic") |> head()
con |> tbl("BcChkERR") |> group_by(type) |> summarise(n=n()) |> arrange(desc(n))
RSQLite::dbDisconnect(con)
# use on.exit to disconnect in functions
}
